FIX/ Grouped Jobs/ New grouped keyed jobs dont crash ExJob
ExJob has has the ability to serialize execution of a group
of jobs mapped to a particular key.

The ages map is used to select which is the next key to work
with. This map is not correctly cleaned up when there are no more
pending jobs for a particular key

The result: when a job with a completely new key is added after jobs
with another key have been executed, the older key selected, even though
there are no pending jobs associated with that key, causing a crash.